quick key bug fix
I was casting the key to a byte instead of a short for some reason, so it was wrapping the keys around. Fixed now!

+++ src/main/java/jgl/Keyboard.java
@@ -74,7 +74,7 @@ public final class Keyboard {
             if (key < 0 || key > GLFW_KEY_LAST)
                 return;
 
-            currentKey = (byte) key;
+            currentKey = (short) key;
             modifierState = (byte) mods;
 
             switch (action) {
